+++ client/components/category/categoryDetail/categoryDetailForm.spec.js
@@ -1,5 +1,5 @@
 import { shallow } from 'enzyme';
-import TextField from '../common/TextField.jsx';
+import TextField from '../../common/TextField.jsx';
 
 import React from 'react';
 import CategoryDetailForm from './CategoryDetailForm.jsx';