+++ client/components/common/searchBar.spec.js
@@ -39,15 +39,15 @@ describe('<SearchBar />', () => {
     };
 
     const redirectSpy = sinon.spy();
-    const browserHistory = require('react-router').browserHistory;
-    browserHistory.push = redirectSpy;
+    const history = require('../../utilities/history').default;
+    history.push = redirectSpy;
 
     const expected = true;
     shallow(<SearchBar.WrappedComponent {...updatedProps} />)
       .find('.logout-link')
       .simulate('click');
 
-    const actual = redirectSpy.calledWith('login');
+    const actual = redirectSpy.calledWith('/login');
     expect(actual).to.equal(expected);
   });
 
@@ -61,8 +61,8 @@ describe('<SearchBar />', () => {
     };
 
     const redirectSpy = sinon.spy();
-    const browserHistory = require('react-router').browserHistory;
-    browserHistory.push = redirectSpy;
+    const history = require('../../utilities/history').default;
+    history.push = redirectSpy;
 
     const expected = true;
     shallow(<SearchBar.WrappedComponent {...updatedProps} />)